Move the selections out of WaylandState
`wayland.rs` was 2665 lines and one struct with 50 fields whose only common
theme was the word "Wayland". The clipboard and the primary selection are
the most detachable part: ten fields, five handler impls, the prefetch path
and the pipe reader, none of it touched from outside the module.

They move to `platform::selections` whole. The `delegate_*` macros need the
handler impls on `WaylandState`, but an impl can live in any module of the
crate, so the code travels with the state instead of only the fields moving
— which is the point, otherwise the file stays the same size.

Two behavioural details worth naming, both preserved: the devices are bound
to the seat when the keyboard capability arrives, not at startup with their
managers (now `Selections::attach_devices`), and the generation counters
still drop a read that a newer offer has overtaken.

wayland.rs 2665 -> 2211 lines, 50 -> 35 fields.
